Add c# client to aspnet core ChatSample (#338)
* Add c# client to aspnet core ChatSample

* Fix build failure
