+++ AzureSignalR.sln
@@ -12,8 +12,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{C4BC
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.SignalR", "src\Microsoft.Azure.SignalR\Microsoft.Azure.SignalR.csproj", "{7B124E40-36AC-45CC-9467-F6A50D5FC149}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample", "samples\ChatSample\ChatSample.csproj", "{B23C2F84-B0B8-47D1-A8D2-E524F6511215}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{36024D6C-A549-4269-B11B-4BFC5F08BD11}"
 	ProjectSection(SolutionItems) = preProject
 		build\dependencies.props = build\dependencies.props
@@ -56,6 +54,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.SignalR.Man
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.SignalR.TestUtility", "test\Microsoft.Azure.SignalR.TestUtility\Microsoft.Azure.SignalR.TestUtility.csproj", "{4F89E9C0-C7A0-40D7-B6E8-406564862975}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ChatSample", "ChatSample", "{C965ED06-6A17-4329-B3C6-811830F4F4ED}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample", "samples\ChatSample\ChatSample\ChatSample.csproj", "{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample.CSharpClient", "samples\ChatSample\ChatSample.CSharpClient\ChatSample.CSharpClient.csproj", "{C51AA61E-5210-4545-B49E-61255C136B83}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -66,10 +70,6 @@ Global
 		{7B124E40-36AC-45CC-9467-F6A50D5FC149}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7B124E40-36AC-45CC-9467-F6A50D5FC149}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7B124E40-36AC-45CC-9467-F6A50D5FC149}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B23C2F84-B0B8-47D1-A8D2-E524F6511215}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B23C2F84-B0B8-47D1-A8D2-E524F6511215}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B23C2F84-B0B8-47D1-A8D2-E524F6511215}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B23C2F84-B0B8-47D1-A8D2-E524F6511215}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2EBFCD26-6E31-4855-ABF5-0A82E78F87C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2EBFCD26-6E31-4855-ABF5-0A82E78F87C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2EBFCD26-6E31-4855-ABF5-0A82E78F87C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -122,13 +122,20 @@ Global
 		{4F89E9C0-C7A0-40D7-B6E8-406564862975}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4F89E9C0-C7A0-40D7-B6E8-406564862975}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4F89E9C0-C7A0-40D7-B6E8-406564862975}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C51AA61E-5210-4545-B49E-61255C136B83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C51AA61E-5210-4545-B49E-61255C136B83}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C51AA61E-5210-4545-B49E-61255C136B83}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C51AA61E-5210-4545-B49E-61255C136B83}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{7B124E40-36AC-45CC-9467-F6A50D5FC149} = {DA69F624-5398-4884-87E4-B816698CDE65}
-		{B23C2F84-B0B8-47D1-A8D2-E524F6511215} = {C4BC9889-B49F-41B6-806B-F84941B2549B}
 		{2EBFCD26-6E31-4855-ABF5-0A82E78F87C7} = {DA69F624-5398-4884-87E4-B816698CDE65}
 		{764DCFF5-D28D-4CEF-B186-B7A62FBB3E98} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
 		{E2094C87-C0F2-471D-A399-0B78CC32C1AA} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
@@ -143,6 +150,9 @@ Global
 		{8DC103BA-7404-479F-BF18-90EF3EC3CEA9} = {DA69F624-5398-4884-87E4-B816698CDE65}
 		{493CE126-D147-434E-ACF8-675E89D0A981} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
 		{4F89E9C0-C7A0-40D7-B6E8-406564862975} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
+		{C965ED06-6A17-4329-B3C6-811830F4F4ED} = {C4BC9889-B49F-41B6-806B-F84941B2549B}
+		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9} = {C965ED06-6A17-4329-B3C6-811830F4F4ED}
+		{C51AA61E-5210-4545-B49E-61255C136B83} = {C965ED06-6A17-4329-B3C6-811830F4F4ED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7945A4E4-ACDB-4F6E-95CA-6AC6E7C2CD59}
