+++ samples/ChatSample/ChatSample/Hub/NotificationHub.cs
