+++ samples/ChatSample/ChatSample/Properties/launchSettings.json
