+++ samples/ChatSample/ChatSample/wwwroot/favicon.ico
