+++ samples/ChatSample/ChatSample/wwwroot/scripts/bootstrap.js
