+++ samples/ChatSample/ChatSample/wwwroot/scripts/jquery-1.10.2.js
