+++ samples/ChatSample/ChatSample/wwwroot/scripts/msgpack5.js
