+++ samples/ChatSample/ChatSample/wwwroot/scripts/signalr-protocol-msgpack.js
